Check for "known" seed entropy, and add watermark to SLIP-39 output o Start by reorganizing, and add tests for some known entropy o Begin to add copyright headers
