+++ GNUmakefile
@@ -41,7 +41,7 @@ else
 endif
 
 # To see all pytest output, uncomment --capture=no
-PYTESTOPTS	= -vv --doctest-modules --capture=no --log-cli-level=INFO
+PYTESTOPTS	= -vv --doctest-modules --capture=no # --log-cli-level=INFO
 
 PY3TEST		= $(PY3) -m pytest $(PYTESTOPTS)
