+++ slip39/api.py
@@ -1,3 +1,20 @@
+
+#
+# Python-slip39 -- Ethereum SLIP-39 Account Generation and Recovery
+#
+# Copyright (c) 2022, Dominion Research & Development Corp.
+#
+# Python-slip39 is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.  It is also available under alternative (eg. Commercial) licenses, at
+# your option.  See the LICENSE file at the top of the source tree.
+#
+# Python-slip39 is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+
 import base58
 import codecs
 import hashlib
@@ -226,7 +243,7 @@ class XRPHDWallet( hdwallet.HDWallet ):
 
 class Account:
     """A Cryptocurrency "Account" / Wallet, based on a variety of underlying Python crypto-asset
-    support modules.  Presently, only meherett/python-hdwallet is used
+    support modules.  Presently, only meherett/python-hdwallet is used.
 
     An appropriate hdwallet-like wrapper is built, for any crypto-asset supported using another
     module.  The required hdwallet API calls are:
@@ -441,11 +458,15 @@ class Account:
 
     @property
     def address( self ):
-        if self.format == "legacy":
+        """Returns the 1..., 3... or bc1... address, depending on whether format is legacy, segwit or bech32"""
+        return self.formatted_address()
+
+    def formatted_address( self, format=None ):
+        if ( format or self.format or '' ).lower() == "legacy":
             return self.legacy_address()
-        elif self.format == "segwit":
+        elif ( format or self.format or '' ).lower() == "segwit":
             return self.segwit_address()
-        elif self.format == "bech32":
+        elif ( format or self.format or '' ).lower() == "bech32":
             return self.bech32_address()
         raise ValueError( f"Unknown addresses semantic: {self.format}" )
 
@@ -496,8 +517,17 @@ class Account:
         return self.hdwallet.private_key()
 
     @property
+    def xkey( self ):
+        return self.hdwallet.xprivate_key()
+
+    @property
     def pubkey( self ):
         return self.hdwallet.public_key()
+
+    @property
+    def xpubkey( self ):
+        """Returns the xpub, ypub or zpub, depending on whether format is legacy, segwit or bech32"""
+        return self.hdwallet.xpublic_key()
 
     def from_private_key( self, private_key ):
         self.hdwallet.from_private_key( private_key )
@@ -976,8 +1006,8 @@ def accountgroups(
     yield from zip( *[
         accounts(
             master_secret	= master_secret,
-            paths		= paths,
             crypto		= crypto,
+            paths		= paths,
             allow_unbounded	= allow_unbounded,
         )
         for crypto,paths in cryptopaths_parser( cryptopaths )
