+++ slip39/known/main.py
@@ -1,0 +1,14 @@
+
+from ..recovery.main	import main as recovery_main
+
+def main( argv=None ):
+    """Produce all "known" seeds, both in raw and BIP-39 encoded form (w/ no password).  We'll get
+    these from BIP-39 and SLIP-39 test vectors.
+
+    Finds all .bip39 and .slip39 files
+    
+    """
+
+    #for bip39 in 
+    
+    return 0
