+++ slip39/recovery_test.csv
@@ -1,0 +1,7 @@
+entropy, passphrase, using_bip39, path, address
+zoo zoo zoo zoo zoo zoo zoo zoo zoo zoo zoo wrong,,true,,bc1qk0a9hr7wjfxeenz9nwenw9flhq0tmsf6vsgnn2
+zoo zoo zoo zoo zoo zoo zoo zoo zoo zoo zoo wrong,,,,bc1q9yscq3l2yfxlvnlk3cszpqefparrv7tk24u6pl
+fruit wave dwarf banana earth journey tattoo true farm silk olive fence,banana,true,,17rxURoF96VhmkcEGCj5LNQkmN9HVhWb7F
+000102030405060708090a0b0c0d0e0f,,,m/0'/1,xpub6ASuArnXKPbfEwhqN6e3mwBcDTgzisQN1wXN9BJcM47sSikHjJf3UFHKkNAWbWMiGj7Wf5uMash7SyYq527Hqck2AxYysAA7xmALppuCkwQ
+fffcf9f6f3f0edeae7e4e1dedbd8d5d2cfccc9c6c3c0bdbab7b4b1aeaba8a5a29f9c999693908d8a8784817e7b7875726f6c696663605d5a5754514e4b484542,,,m/0,xpub69H7F5d8KSRgmmdJg2KhpAK8SR3DjMwAdkxj3ZuxV27CprR9LgpeyGmXUbC6wb7ERfvrnKZjXoUmmDznezpbZb7ap6r1D3tgFxHmwMkQTPH
+fffcf9f6f3f0edeae7e4e1dedbd8d5d2cfccc9c6c3c0bdbab7b4b1aeaba8a5a29f9c999693908d8a8784817e7b7875726f6c696663605d5a5754514e4b484542,,,m/0/2147483647'/1,xpub6DF8uhdarytz3FWdA8TvFSvvAh8dP3283MY7p2V4SeE2wyWmG5mg5EwVvmdMVCQcoNJxGoWaU9DCWh89LojfZ537wTfunKau47EL2dhHKon
