+++ slip39/version.py
@@ -1,2 +1,1 @@
-__version_info__		= ( 9, 0, 4 )
-__version__			= '.'.join( map( str, __version_info__ ))
+__version_info__		= ( 9, 1, 0 )
